Code refactor

+++ src/CartConditions/Tax.php
@@ -31,7 +31,7 @@ class Tax extends CartCondition
     {
         $this->taxMode = (bool)setting('tax_mode', 1);
         $this->taxInclusive = !((bool)setting('tax_menu_price', 1));
-        $this->taxRate = $this->taxRateLabel = setting('tax_percentage', 0);
+        $this->taxRate = $this->taxRateLabel = (int)setting('tax_percentage', 0);
         if ($this->taxInclusive) {
             $this->taxRate /= (100 + $this->taxRate) / 100;
         }
@@ -55,7 +55,7 @@ class Tax extends CartCondition
             [
                 'value' => "+{$this->taxRate}%",
                 'inclusive' => $this->taxInclusive,
-                'valuePrecision' => $precision,
+                'valuePrecision' => (int)$precision,
             ],
         ];
     }
